feat: resolve local package by name
This PR ensures that rules implemented locally are also resolved locally
within the package.

To do that we check for a local `package.json`. If it exists we use the
package name to normalize required configs and rules.

Adjustment must be done both when resolving rules as well as when
compiling the linter configuration.